Remove #include "statements.h" from visitors.h
