+++ taichi/analysis/verify.cpp
@@ -3,6 +3,7 @@
 
 #include "taichi/ir/ir.h"
 #include "taichi/ir/analysis.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/ir/transforms.h"
