+++ taichi/ir/basic_stmt_visitor.cpp
@@ -1,5 +1,6 @@
-#include "taichi/ir/visitors.h"
 #include "taichi/ir/frontend_ir.h"
+#include "taichi/ir/statements.h"
+#include "taichi/ir/visitors.h"
 
 TLANG_NAMESPACE_BEGIN
