+++ taichi/ir/visitors.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "taichi/ir/ir.h"
-#include "taichi/ir/statements.h"
+#include "taichi/ir/statement_names.h"
 
 TLANG_NAMESPACE_BEGIN
