+++ taichi/transforms/die.cpp
@@ -1,6 +1,7 @@
 // Dead Instruction Elimination
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
 #include <unordered_set>
