+++ taichi/transforms/ir_printer.cpp
@@ -1,6 +1,7 @@
 // The IRPrinter prints the IR in a human-readable format
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
 #include "taichi/ir/frontend_ir.h"
