+++ taichi/transforms/loop_vectorize.cpp
@@ -1,6 +1,7 @@
 // The loop vectorizer
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
