+++ taichi/transforms/make_thread_local.cpp
@@ -5,6 +5,7 @@
 
 #include "taichi/ir/analysis.h"
 #include "taichi/ir/ir.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
