+++ taichi/transforms/unreachable_code_elimination.cpp
@@ -1,6 +1,7 @@
 #include "taichi/ir/ir.h"
-#include "taichi/ir/visitors.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
+#include "taichi/ir/visitors.h"
 
 TLANG_NAMESPACE_BEGIN
