+++ taichi/transforms/vector_split.cpp
@@ -1,6 +1,7 @@
 // Split vectors wider than machine vector width into multiple vectors
 
 #include "taichi/ir/ir.h"
+#include "taichi/ir/statements.h"
 #include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
 #include <algorithm>
