+++ taichi/transforms/whole_kernel_cse.cpp
@@ -1,6 +1,7 @@
 #include "taichi/ir/ir.h"
-#include "taichi/ir/transforms.h"
 #include "taichi/ir/analysis.h"
+#include "taichi/ir/statements.h"
+#include "taichi/ir/transforms.h"
 #include "taichi/ir/visitors.h"
 #include <typeindex>
